fix(backend,db): resolve all 23 findings from the full backend + database audit
SECURITY (high):
- platform_audit had NO RLS: qvm_app in any tenant context could read/insert/delete the cross-
  workspace impersonation + admin ledger (proved live). 0034 enables + FORCEs RLS with SELECT
  internal-only and INSERT internal-or-own-tenant, and NO update/delete policy → append-only.
  Root cause was systemic: 0001 applies policies in a ONE-SHOT loop while 0002 grants qvm_app DML
  on all FUTURE tables, so any later table is born open. Added scripts/verify-rls.ts +
  `pnpm --filter @qvm/api db:verify` asserting every table has RLS, policies, FORCE and a
  tenant-scoped policy (86/86 pass on local + prod); corrected the misleading 0001/platform_audit
  comments that claimed future coverage.

CORRECTNESS (high):
- The two quote-write paths had DIVERGED: the token path used onConflictDoNothing (a vendor
  re-quote via the emailed link was silently dropped while still reporting success) vs the portal
  path upserting. Both now call one shared VendorRfqService.writeQuoteItems (upsert semantics),
  and submitQuoteSchema is defined once and re-exported.
- Duplicate line items in ONE payload bypassed the over-delivery / over-return / payment-allocation
  caps (each line checked against the stored total independently). All three now aggregate per key
  BEFORE validating; delivery also flips item status on the aggregated qty.
- purchase_orders.invoice_amount had no writer, so vendor statements always read 0.00 and the
  allocation cap never engaged: added POST /purchase-orders/:poId/invoice (internal, set-once).
- Idempotency had no DB backstop: added unique invoices(order_id) + purchase_orders(order_id,
  vendor_id) with 23505 → 400 mapping, so concurrent double-issue/double-create cannot slip past
  the check-then-insert.
- orders.confirm() never advanced rfq_vendors to confirmed → the vendor portal "won" KPI was a
  dead filter; it now advances the winning vendors in the same transaction.
- org.createBranch accepted orderCategory "insurance" which the pg enum lacks (guaranteed 500).

CONSISTENCY / DUPLICATION:
- New shared helpers: resolveCounterparty/requireCounterparty (the counterparty lookup was
  repeated 5 ways), targetTenant, assertRfqNotConfirmed, requireTenantCtx/openCtx (rfq + orders
  controllers no longer hand-roll ctx and now forward environment + impersonatorId).
- 23505 in counterparty.approve now maps to 409 like every other unique-violation.
- vendors.controller platform-only writes use ctxOpen so dto.tenantId targeting actually works.
- Removed dead isInternalRole/PLATFORM_ROLES; documented the reserved app.impersonator_id GUC.
- count() aggregates returned to the API are cast ::int (were leaking as strings).

SCHEMA / MIGRATIONS:
- 0035 refreshes the drizzle snapshot (stalled at 0029 because 0030-0034 were hand-written);
  generated SQL was additions-only with zero DROPs, confirming .ts ↔ DB are in lockstep, so it is
  a documented no-op. `drizzle-kit generate` now reports "No schema changes".
- Declared the two DB-only unique indexes in .ts; renamed the counterparty_submissions FK to the
  drizzle convention; seed now writes rfq_items.status_id and mirrors confirm() exactly.

Verified: 46/46 local regression + 16/16 prod suite; targeted proofs for every fix (dup-line
over-delivery 400, invoice writer 260.00 → statement invoiced 260.00/unpaid, double-PO 400,
set-once invoice 400, won=1, platform_audit rows invisible to a tenant session, RLS 86/86).
Applied to local + prod (36 migrations recorded on both); typecheck clean.

+++ apps/api/src/modules/admin/workspaces-admin.service.ts
@@ -30,10 +30,10 @@ export class WorkspacesAdminService {
     const rows = await this.dbService.withContext(INTERNAL, (tx) =>
       tx.execute(sql`
         select t.id, t.slug, t.name, t.is_sandbox, t.is_active, t.created_at,
-          (select count(*) from tenant_workshops tw join workshop_branches wb on wb.workshop_id = tw.workshop_id
+          (select count(*)::int from tenant_workshops tw join workshop_branches wb on wb.workshop_id = tw.workshop_id
             where tw.tenant_id = t.id and tw.status <> 'archived') as branches,
-          (select count(*) from tenant_vendors tv where tv.tenant_id = t.id and tv.status = 'active') as vendors,
-          (select count(*) from tenant_memberships m where m.tenant_id = t.id and m.is_active = true) as users
+          (select count(*)::int from tenant_vendors tv where tv.tenant_id = t.id and tv.status = 'active') as vendors,
+          (select count(*)::int from tenant_memberships m where m.tenant_id = t.id and m.is_active = true) as users
         from tenants t order by t.created_at desc`),
     );
     return { count: rows.length, workspaces: rows };
@@ -87,7 +87,7 @@ export class WorkspacesAdminService {
 
       const workshops = await tx.execute(sql`
         select w.id, w.name, w.tax_number,
-          (select count(*) from workshop_branches wb where wb.workshop_id = w.id) as branches
+          (select count(*)::int from workshop_branches wb where wb.workshop_id = w.id) as branches
         from tenant_workshops tw join workshops w on w.id = tw.workshop_id
         where tw.tenant_id = ${id}::uuid and tw.status <> 'archived' order by w.name`);
 
@@ -100,14 +100,14 @@ export class WorkspacesAdminService {
       // ---- operational sections (environment-scoped) ----
       const rfqs = await tx.execute(sql`
         select r.id, r.order_number, r.plate_number, s.label_en as status,
-               (select count(*) from rfq_items i where i.rfq_id = r.id) as items
+               (select count(*)::int from rfq_items i where i.rfq_id = r.id) as items
         from rfqs r left join item_statuses s on s.id = r.status_id
         where r.tenant_id = ${id}::uuid and r.environment = ${environment}
         order by r.created_at desc limit 25`);
 
       const orders = await tx.execute(sql`
         select o.id, o.order_number, s.label_en as status,
-               (select count(*) from order_items oi where oi.order_id = o.id) as items
+               (select count(*)::int from order_items oi where oi.order_id = o.id) as items
         from orders o left join item_statuses s on s.id = o.status_id
         where o.tenant_id = ${id}::uuid and o.environment = ${environment}
         order by o.created_at desc limit 25`);
